fix(misc): pass-4 minor cleanup (E1-E4)
E1: MultimodalCollator now checks for empty batch before checking
require_images. Empty input is the more fundamental failure -- there is
nothing to process at all, so the pixel_values state is irrelevant.

E2: PackedCollator docstring corrected from "long" to "int8" for the
attention_mask dtype, matching the actual implementation.

E3: test_config_rejects_zero_lora_r_for_lora parametrized to cover
both "lora" and "qlora" methods. The validator applies to both but
only "lora" was tested.

E4: test_pack_unpack_empty_tensor_roundtrip parametrized to also cover
(out_features=4, in_features=0), verifying that pack_int4/unpack_int4
handle a weight with rows but no columns without crashing.

+++ mud_puppy/packing.py
@@ -30,7 +30,7 @@ class PackedCollator:
     emitting:
 
     - ``input_ids``: shape (batch, seq)
-    - ``attention_mask``: shape (batch, seq, seq) -- block-diagonal, long
+    - ``attention_mask``: shape (batch, seq, seq) -- block-diagonal, int8
     - ``labels``: shape (batch, seq) -- preserves -100 from source
     - ``position_ids``: shape (batch, seq) -- resets per segment
